toolkit/mozapps/update/tests/chrome/test_0097_restartNotification_stagedServiceBackground.xul

<?xml version="1.0"?>
<!--
/* Any copyright is dedicated to the Public Domain.
 * http://creativecommons.org/publicdomain/zero/1.0/
 */
-->

<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<window title="Update Wizard pages: restart notification staged with service"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="runTestDefault();">
<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
<script type="application/javascript"
        src="utils.js"/>

<script type="application/javascript">
<![CDATA[

const TESTS = [ {
  pageid: PAGEID_FINISHED_BKGRD,
  buttonClick: "extra1"
} ];

gUseTestUpdater = true;

function runTest() {
  debugDump("entering");

  let patchProps = {state: STATE_APPLIED_SVC};
  let patches = getLocalPatchString(patchProps);
  let updates = getLocalUpdateString({}, patches);
  writeUpdatesToXMLFile(getLocalUpdatesXMLString(updates), true);
  writeStatusFile(STATE_APPLIED_SVC);

  Services.prefs.setIntPref(PREF_APP_UPDATE_PROMPTWAITTIME, 1);
  Services.prefs.setBoolPref(PREF_APP_UPDATE_STAGING_ENABLED, true);
  Services.prefs.setBoolPref(PREF_APP_UPDATE_SERVICE_ENABLED, true);

  reloadUpdateManagerData();

  is(gUpdateManager.activeUpdate.state, STATE_APPLIED_SVC,
     "The active update should have a state of " + STATE_APPLIED_SVC);

  ok(gUpdateManager.activeUpdate.promptWaitTime == 1, "Checking that the " +
     "update's promptWaitTime attribute value was set from the " +
     PREF_APP_UPDATE_PROMPTWAITTIME + " preference");

  gUpdateManager.refreshUpdateStatus();
}

]]>
</script>

<body xmlns="http://www.w3.org/1999/xhtml">
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test"></pre>
</body>
</window>
